Fixed logic for updating message file in template
Logic now work by searching in the provided file by git to prepare for
comments, ussually git adds some comments to the prepared message, is
this not the case the logic needs to be re-though again.

After first line with comment (#) is found the message is plsit in two
part, before comments starts, which is the original message, and the
rest of the message containing ussually some comments plus diff of the
changes done in the commit.

From there templates can work with those two variables to modify the
message. So far use cases are to prepend some text or to append some text
to the message

Changes:
* Ignore testing/resources direectory in build/tests
* Start using second parameter from prepare-commit-msg hook as per docs:
  https://git-scm.com/docs/githooks#_prepare_commit_msg, this allow for
  ignore scenarios where it is not wanted to modify the commit message
* Adds more tests, file expected and actual files. More close to whats
  expected in regular git interaction.
* Updates logic and templates for updating the message.